backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pypdf2==3.0.1
ollama==0.3.3
python-dotenv==1.0.1
